refactor: 완료페이지 API HMGArea 컴포넌트로 이동
