+++ frontend/src/pages/EstimationPage/index.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 import { useReactToPrint } from 'react-to-print';
 import { useData, TotalPrice } from '../../utils/Context';
 import * as S from './style';
@@ -8,7 +8,6 @@ import Info from './Info';
 import Detail from './Detail';
 import HMGArea from './HMGArea';
 import Footer from './Footer';
-import useFetch from '../../hooks/useFetch';
 
 function Estimation() {
   const data = useData();
@@ -18,31 +17,6 @@ function Estimation() {
     content: () => pdfRef.current,
     documentTitle: '내 차 만들기_견적서',
   });
-  const fetchedPostData = useFetch('estimates', {
-    method: 'POST',
-    body: {
-      detailTrimId: data.modelType.detailTrimId,
-      exteriorColorCode: data.exteriorColor.code,
-      interiorColorCode: data.interiorColor.code,
-      selectOptionOrPackageIds: data.optionPicker.chosenOptions,
-    },
-  });
-  const fetchedGetData = useFetch(`estimates/distribution?trimId=${data.trim.id}`);
-
-  useEffect(() => {
-    if (!fetchedPostData || !fetchedGetData || data.estimation.isFetch || data.page !== 6) return;
-    data.setTrimState((prevState) => ({
-      ...prevState,
-      page: 6,
-      estimation: {
-        ...prevState.estimation,
-        isPost: true,
-        id: fetchedPostData,
-        isFetch: true,
-        averagePrice: fetchedGetData,
-      },
-    }));
-  }, [data, fetchedGetData, fetchedPostData]);
 
   return (
     <S.Estimation>
